2dcontext/scroll/scrollPathIntoView.verticalLR.html

<!DOCTYPE html>
<html style="writing-mode: vertical-lr;">
<head>
<meta charset="UTF-8">
<title>scrollPathIntoView in vertical-lr writing mode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
</head>
<body>
<div id="log"></div>
<canvas id="c" width="300" height="150" style="position: absolute; left: 100px; top: 200px;"></canvas>
<div style="width: 200vw; height: 200vh;"></div>
<script>
var t = async_test('');
_addTest(function (canvas, ctx) {
    window.scrollTo(0, 0);
    ctx.beginPath();
    ctx.rect(10, 20, 40, 80);
    ctx.scrollPathIntoView();
    var rect = c.getBoundingClientRect();
    assert_approx_equals(rect.left, -10, 0.5, 'The canvas left edge is on 10px left of the viewport left edge');
    assert_approx_equals(rect.top, 200, 0.5, 'The canvas top edge is 200px below the viewport top edge');
});
</script>
</body>
</html>
